PYTH/task-service/requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
pytest==7.3.1
python-dotenv==1.0.0
marshmallow==3.19.0
psycopg2-binary==2.9.6
flask-cors==3.0.10
gunicorn==20.1.0
